Remove dead DEX liquidity fallback and challenger-probe code
The inline discovery fallbacks were disabled in the scoring lane, leaving
several DexScreener/CoinGecko-tickers crawlers and their always-zero
diagnostics reachable only from tests. Three related cleanups:

A2 — delete the dead fallback fetchers (fetchDsFallbackPools,
fetchCgTickersFallback, getDsFallbackTargets, getCgTickersFallbackTargets,
needsOrderbookFallback). getFallbackTargets and its weak-coverage
predicates stay: orchestrator-phases/fallback.ts still uses them. The live
"CoinGecko tickers shared helpers" coverage moved to its own
coingecko-tickers-shared test before fetch-fallbacks.test.ts was deleted,
and dex-liquidity-fallbacks.test.ts keeps the getFallbackTargets cases plus
a new one pinning the measured-balance-only branch that the deleted file
was the last to assert.

A4 — drop detectDexPriceChallengerTableState and the missing-table branches
in challenger load/publish. Both challenger tables ship in
0000_baseline.sql and migrations run before the new Worker is live, so the
state is unreachable. Per-query isMissingTableError catches and the live
challenger-legacy degraded-producer lane are untouched; the challenger
tests now express "tables absent" through those catches. sync-dex-liquidity
tests mock the challenger publisher, whose real D1 change-count assert the
removed probe used to short-circuit under the minimal db mock.

A5 — collapse the always-zero fallback diagnostics: one getFallbackTargets
call in runFallbackCrawlerPhase (it was called twice with identical args,
so coverageRecoveredCoins was structurally always 0) and drop
coverageRecoveredCoins/dsFallbackCoins/cgTickerFallbackCoins from the phase
result and the orchestrator/analysis plumbing.
weakCoverageCoinsBeforeFallback is retained. The metadata schema fields
are optional and the status reader null-guards, so no consumer changes.
