+++ docs/dex-liquidity.md
@@ -27,7 +27,7 @@ When DeFiLlama Protocols is unavailable, protocol TVL caps cannot be computed re
 
 When DeFiLlama Yields is unavailable, the cron treats value/coverage guard failures as source-incomplete degradation instead of throwing before metadata can be written. The run returns `degraded`, skips persistence, and keeps the last successful public dataset authoritative until a source-complete run recovers.
 
-Run metadata now includes `failedSources`, `fallbackMode` signals, staged-pool merge counters (`stagedPoolsMerged`, `stagedPoolsSkipped`, `stagedPoolsSkippedByExactIdentity`, `stagedPoolsSkippedByUniqueDerivedIdentity`, `stagedPoolsSkippedByOptionalWildcardIdentity`, `stagedPoolsSkippedByAuthoritativeProtocol`), challenger publish counters, persistence skip state, inactive tracked-asset skip counts, publication-generation diagnostics (`generationId`, expected/candidate/current row counts), and detailed `sourceCoverage` values (`currentCoverage`, `previousCoverage`, `minExpectedCoverage`, `nearCoverageGuard`, `currentGlobalTvl`, `previousGlobalTvl`, `minExpectedGlobalTvl`, `valueBaselineSource`, `valueBaselineGlobalTvl`, `ignoredPersistedGlobalTvl`, `nearValueGuard`, `currentTop10CoveredTvl`, `previousTop10CoveredTvl`, `currentTop10GuardTvl`, `previousTop10GuardTvl`, `nearMajorCoverageGuard`, `currentCoverageClasses`, `previousCoverageClasses`, `priceObservationCoins`, `dsFallbackCoins`, `cgTickerFallbackCoins`).
+Run metadata now includes `failedSources`, `fallbackMode` signals, staged-pool merge counters (`stagedPoolsMerged`, `stagedPoolsSkipped`, `stagedPoolsSkippedByExactIdentity`, `stagedPoolsSkippedByUniqueDerivedIdentity`, `stagedPoolsSkippedByOptionalWildcardIdentity`, `stagedPoolsSkippedByAuthoritativeProtocol`), challenger publish counters, persistence skip state, inactive tracked-asset skip counts, publication-generation diagnostics (`generationId`, expected/candidate/current row counts), and detailed `sourceCoverage` values (`currentCoverage`, `previousCoverage`, `minExpectedCoverage`, `nearCoverageGuard`, `currentGlobalTvl`, `previousGlobalTvl`, `minExpectedGlobalTvl`, `valueBaselineSource`, `valueBaselineGlobalTvl`, `ignoredPersistedGlobalTvl`, `nearValueGuard`, `currentTop10CoveredTvl`, `previousTop10CoveredTvl`, `currentTop10GuardTvl`, `previousTop10GuardTvl`, `nearMajorCoverageGuard`, `currentCoverageClasses`, `previousCoverageClasses`, `priceObservationCoins`, `weakCoverageCoins`).
 
 Current-row publication is generation-gated and active-set scoped. The cron may keep historical rows for inactive tracked assets, but the public `dex_liquidity` current table is rewritten from the current active tracked universe plus the `__global__` aggregate only. Candidate rows are written to `dex_liquidity_run_rows`, the expected active row count is validated, and only then is the candidate generation mirrored into the public table and marked `dex_liquidity_publication_generations.state = 'published'`. Readers accept legacy rows with no generation id during rollout, but generation-tagged rows are consumed only when their generation is published.
