+++ worker/src/cron/__tests__/dex-liquidity-fallbacks.test.ts
@@ -1,46 +1,7 @@
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { mockCircuitBreaker, mockCircuitOutcomeRecord } from "../../test-helpers/cron";
+import { describe, expect, it } from "vitest";
 
-vi.mock("../../lib/circuit-breaker", () => mockCircuitBreaker());
-
-vi.mock("../../lib/dexscreener", async () => {
-  const actual = await vi.importActual<typeof import("../../lib/dexscreener")>("../../lib/dexscreener");
-  return {
-    ...actual,
-    fetchDsTokenPoolsWithStatus: vi.fn(async () => ({ ok: true, pairs: [] })),
-    dsRateLimit: vi.fn(async () => {}),
-  };
-});
-
-vi.mock("../../lib/cron-logger", () => ({
-  logCronEvent: vi.fn(async () => {}),
-}));
-
-import { fetchDsFallbackPools, getFallbackTargets } from "../dex-liquidity/fetch-fallbacks";
-import { createKnownPoolIdentityIndex } from "../dex-liquidity/pool-identity";
+import { getFallbackTargets } from "../dex-liquidity/fetch-fallbacks";
 import { initMetrics } from "../dex-liquidity/pool-helpers";
-import { shouldAttemptFetch, recordOutcome } from "../../lib/circuit-breaker";
-import { fetchDsTokenPoolsWithStatus } from "../../lib/dexscreener";
-import { CIRCUIT_SOURCE } from "../../lib/constants";
-import { logCronEvent } from "../../lib/cron-logger";
-
-function createMockDb(): D1Database {
-  return {
-    prepare: () => ({
-      bind: () => ({
-        all: async () => ({ results: [] }),
-        first: async () => null,
-        run: async () => ({ success: true, meta: {} }),
-      }),
-      all: async () => ({ results: [] }),
-      first: async () => null,
-      run: async () => ({ success: true, meta: {} }),
-    }),
-    batch: async () => [],
-    exec: async () => ({ count: 0, duration: 0 }),
-    dump: async () => new ArrayBuffer(0),
-  } as unknown as D1Database;
-}
 
 describe("getFallbackTargets", () => {
   it("targets coins with zero pools, missing dex price observations, or weak partial coverage", () => {
@@ -79,6 +40,31 @@ describe("getFallbackTargets", () => {
     expect(targetIds.has("dai-makerdao")).toBe(false);
   });
 
+  it("targets a coin whose only weakness is measured-balance coverage", () => {
+    const metrics = new Map<string, ReturnType<typeof initMetrics>>();
+    const weakBalance = initMetrics("dai-makerdao", "DAI");
+    weakBalance.poolCount = 4;
+    weakBalance.totalTvlUsd = 500_000;
+    weakBalance.totalTvlForBalance = 0;
+    metrics.set("dai-makerdao", weakBalance);
+
+    const priceObservations = new Map([
+      [
+        "dai-makerdao",
+        [
+          { price: 1, tvl: 150_000, chain: "ethereum", protocol: "curve" },
+          { price: 1, tvl: 100_000, chain: "base", protocol: "uniswap-v3" },
+        ],
+      ],
+    ]);
+
+    const targetIds = new Set(
+      getFallbackTargets(metrics, priceObservations, { requireTrackedContracts: true }).map((meta) => meta.id),
+    );
+
+    expect(targetIds.has("dai-makerdao")).toBe(true);
+  });
+
   it("can restrict orderbook fallback targets to coins with a geckoId", () => {
     const metrics = new Map<string, ReturnType<typeof initMetrics>>();
     const noGecko = initMetrics("rwausdi-multipli", "rwaUSDi");
@@ -87,129 +73,5 @@ describe("getFallbackTargets", () => {
     const targetIds = new Set(getFallbackTargets(metrics, new Map(), { requireGeckoId: true }).map((meta) => meta.id));
 
     expect(targetIds.has("rwausdi-multipli")).toBe(false);
-  });
-});
-
-describe("fetchDsFallbackPools circuit breaker", () => {
-  beforeEach(() => {
-    vi.mocked(shouldAttemptFetch).mockReset();
-    vi.mocked(recordOutcome).mockReset();
-    vi.mocked(fetchDsTokenPoolsWithStatus).mockReset();
-    vi.mocked(logCronEvent).mockReset();
-    vi.mocked(shouldAttemptFetch).mockResolvedValue(true);
-    vi.mocked(recordOutcome).mockResolvedValue(mockCircuitOutcomeRecord());
-    vi.mocked(fetchDsTokenPoolsWithStatus).mockResolvedValue({ ok: true, pairs: [] });
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it("skips DexScreener when dexscreener-liquidity breaker is open", async () => {
-    vi.mocked(shouldAttemptFetch).mockImplementation(
-      async (_db, source) => source !== CIRCUIT_SOURCE.DEXSCREENER_LIQUIDITY,
-    );
-
-    // Seed metrics so the fallback has targets (zero pools triggers enrichment).
-    const metrics = new Map<string, ReturnType<typeof initMetrics>>();
-    const zeroPools = initMetrics("usdt-tether", "USDT");
-    zeroPools.poolCount = 0;
-    metrics.set("usdt-tether", zeroPools);
-
-    const db = createMockDb();
-    await fetchDsFallbackPools(db, metrics, new Map(), createKnownPoolIdentityIndex());
-
-    expect(fetchDsTokenPoolsWithStatus).not.toHaveBeenCalled();
-    expect(recordOutcome).not.toHaveBeenCalled();
-  });
-
-  it("records success when DexScreener fetch succeeds", async () => {
-    vi.mocked(shouldAttemptFetch).mockResolvedValue(true);
-    vi.mocked(fetchDsTokenPoolsWithStatus).mockResolvedValue({ ok: true, pairs: [] });
-
-    const metrics = new Map<string, ReturnType<typeof initMetrics>>();
-    const zeroPools = initMetrics("usdt-tether", "USDT");
-    zeroPools.poolCount = 0;
-    metrics.set("usdt-tether", zeroPools);
-
-    const db = createMockDb();
-    await fetchDsFallbackPools(db, metrics, new Map(), createKnownPoolIdentityIndex());
-
-    expect(fetchDsTokenPoolsWithStatus).toHaveBeenCalled();
-    expect(recordOutcome).toHaveBeenCalledWith(expect.anything(), CIRCUIT_SOURCE.DEXSCREENER_LIQUIDITY, true);
-    expect(recordOutcome).toHaveBeenCalledTimes(1);
-  });
-
-  it("records one aggregate failure when DexScreener fetches fail", async () => {
-    vi.mocked(shouldAttemptFetch).mockResolvedValue(true);
-    vi.mocked(fetchDsTokenPoolsWithStatus).mockResolvedValue({ ok: false, pairs: [] });
-
-    const metrics = new Map<string, ReturnType<typeof initMetrics>>();
-    const zeroPools = initMetrics("usdt-tether", "USDT");
-    zeroPools.poolCount = 0;
-    metrics.set("usdt-tether", zeroPools);
-
-    const db = createMockDb();
-    await fetchDsFallbackPools(db, metrics, new Map(), createKnownPoolIdentityIndex());
-
-    expect(fetchDsTokenPoolsWithStatus).toHaveBeenCalled();
-    expect(recordOutcome).toHaveBeenCalledWith(expect.anything(), CIRCUIT_SOURCE.DEXSCREENER_LIQUIDITY, false);
-    expect(recordOutcome).toHaveBeenCalledTimes(1);
-  });
-
-  it("logs DexScreener unusable response details for operations triage", async () => {
-    vi.mocked(shouldAttemptFetch).mockResolvedValue(true);
-    vi.mocked(fetchDsTokenPoolsWithStatus).mockResolvedValue({
-      ok: false,
-      pairs: [],
-      status: 429,
-      contentType: "text/html",
-      error: "HTTP 429 for https://api.dexscreener.com/tokens/v1/ethereum/0xabc; body starts with: rate limited",
-    });
-
-    const metrics = new Map<string, ReturnType<typeof initMetrics>>();
-    const zeroPools = initMetrics("usdt-tether", "USDT");
-    zeroPools.poolCount = 0;
-    metrics.set("usdt-tether", zeroPools);
-
-    const db = createMockDb();
-    await fetchDsFallbackPools(db, metrics, new Map(), createKnownPoolIdentityIndex());
-
-    expect(logCronEvent).toHaveBeenCalledWith(
-      expect.anything(),
-      expect.objectContaining({
-        job: "sync-dex-liquidity",
-        eventType: "dexscreener-fallback-unusable-response",
-        metadata: expect.objectContaining({
-          stablecoinId: "usdt-tether",
-          symbol: "USDT",
-          status: 429,
-          contentType: "text/html",
-          error: expect.stringContaining("HTTP 429"),
-        }),
-      }),
-    );
-  });
-
-  it("tags the between-coins budget-exhausted event with phase metadata", async () => {
-    vi.mocked(shouldAttemptFetch).mockResolvedValue(true);
-
-    const metrics = new Map<string, ReturnType<typeof initMetrics>>();
-    const zeroPools = initMetrics("usdt-tether", "USDT");
-    zeroPools.poolCount = 0;
-    metrics.set("usdt-tether", zeroPools);
-
-    const db = createMockDb();
-    // Deadline already in the past: the per-coin (outer) check fires before any request.
-    await fetchDsFallbackPools(db, metrics, new Map(), createKnownPoolIdentityIndex(), undefined, Date.now() - 1);
-
-    expect(fetchDsTokenPoolsWithStatus).not.toHaveBeenCalled();
-    expect(logCronEvent).toHaveBeenCalledWith(
-      expect.anything(),
-      expect.objectContaining({
-        eventType: "dexscreener-fallback-budget-exhausted",
-        metadata: expect.objectContaining({ phase: "between-coins" }),
-      }),
-    );
   });
 });
