+++ worker/src/cron/dex-liquidity/__tests__/orchestrator-analysis.test.ts
@@ -58,9 +58,6 @@ function makeAnalysisInput(overrides: Partial<Parameters<typeof analyzeDexLiquid
     stagedMergedCount: 0,
     stagedSkippedCount: 0,
     weakCoverageCoinsBeforeFallback: 0,
-    coverageRecoveredCoins: 0,
-    dsFallbackCoins: 0,
-    cgTickerFallbackCoins: 0,
     dlYieldsAvailable: true,
     dlProtocolsAvailable: true,
     directCexOrderbookDepth: null,
